New generic NN script
In place of having to maintain multiple scripts implementing the same neural net but with different numbers of layers, I have created a new script `survey_NN.py` that has the number of hidden layers as a tunable hyperparam. I will later do the same thing for the `survey-kt*` files, or maybe merge them, so the KTAB infusion is an option.