Correct Dev Set Error Calcs
The new CDMP-infused neural network script now correctly computes the model performance metrics on the dev set.  It does this by breaking the dev set into minibatches and applying the trained NN on each batch.  This is clearly suboptimal, but I could find no way around the limitation of using the `batch_size` for indexing into the CDMP "layer".