Automate multiple runs
This commit finalizes the work on `survey_KT.py` to allow it to read model run parameters from a data file.  A single data file can have many model run parameters stored in it.  The `model_input.txt` file has parameters for two NN model runs which are identical except one is a traditional deep network, and the other adds the CDMP infusion.