Jah edits (#14)
* update

* Several changes:
- `survey-synth.py` adds the generating `prng_seed` to the end of the header line of the saved data file
- `survey_KT.py` now reads and prints this seed if extant
- exceptions now catch missing `dataInput` or `modelInput` files
- actual and predicted class labels are decoded from OHE into regular integer labels so that the metrics work correctly

* merged

* Change seed

* Update survey preparation

I realized this was outputting the response as a single nominal categorical column but the model code is expecting one-hot-encoding.  This preparation script has been updated accordingly.

* Preparation script minor edit

* Update hyperparameters
- momentums beta1 and beta2 added for RMSProp and ADAM optimizer
- learning rate decay added

* Create script to generate model inputs for hyperparameter tuning

* More work on tuning scenario generator.

* Update hyperparam scenario generator

Instead of the ill-conceived cartesian grid, this script now just samples randomly from a high-dimensional volume.  The granularity of the sampling and prng seed are user input.  The seed is currently used both to control the scenario generation and written into the scenarios.  This may change.

* Remove old files

* Update modeling script

Modeling script file edited to terminate early if the loss function becomes nan.

* Properly handle nan loss/predictions

Updated the `survey_KT.py` script to terminate early if the loss becomes nan.  Also, nan predictions are cast to 0 for the csv export.

* Update Scenario Volume Generator

Hyperparameter tuning scenario generator now takes as input an int/boolean flag (1/0) of whether or not to use the CDMP.

* Correct logic to equalize batch size for dev set predictins

* Update to use data-relative-sized hidden layer widths:
- `generateScenarioVolume.py` generates hidden layer widths as scaling multiples; also, first layer is hardcoded to be 1.5 times number data input columnbs
-`survey_KT.py` `ReadData` takes hidden layer widths as floats, `RunNN` takes hidden layer widths as float scalers, and multiplies by `num_data_col`

* `generateScenarioVolume.py` prng seed

can now be set to be the same (same seed as generator) for all scenarios, or randomly generated.

* Update `survey_KT.py`

At the top of the loop over scenarios in the main proc, close all open plot windows.

* Update `RunNN` to save `actu_prec.csv` with run-specific name.

* Add scenario name

-`generateScenarioVolume.py` now generates a random 10-char long "name" for scenarios to ease identifying the scenario that relates to a set of outputs.
- `survey_KT.py` modified to read and print the scenario "name"

* Correct a print statement

* Replace sigmoid with softmax

* Add logging

The `survey_KT.py` main procedure now sets up a `logging` object that is use by all functions to print all output (replaced `print` with `logging.info`) to the console and a log file simultaneously.

* Add `scenName` to debug setup commented out code.

* Update Prediction Extraction

- Added new function `predFromProb` to `klib.py` that converts class probabilities to class predictions.
- Modified `RunNN` in `survey_KT.py` to use the new function

* Adjust for nans in `klib.py.predFromProb` before adjusting probabilities.

* Correct `xy_dict` by adding back `pred` after the loss goes nan.